fix(TableInfo): refactor
