+++ src/containers/Tenant/Diagnostics/Overview/TableInfo/TableInfo.tsx
@@ -3,22 +3,13 @@ import React from 'react';
 import {ChevronDown, ChevronUp, Gear} from '@gravity-ui/icons';
 import {ActionTooltip, Button, Disclosure, Flex, Icon} from '@gravity-ui/uikit';
 
-import {YDBDefinitionList} from '../../../../../components/YDBDefinitionList/YDBDefinitionList';
 import {useFeatureFlagsAvailable} from '../../../../../store/reducers/capabilities/hooks';
 import {configsApi} from '../../../../../store/reducers/configs';
 import {operationsApi} from '../../../../../store/reducers/operations';
-import {tablePartitioningApi} from '../../../../../store/reducers/tablePartitioning/tablePartitioning';
 import {EPathType} from '../../../../../types/api/schema';
 import type {TEvDescribeSchemeResult} from '../../../../../types/api/schema';
 import {cn} from '../../../../../utils/cn';
-import createToast from '../../../../../utils/createToast';
-import {useAutoRefreshInterval} from '../../../../../utils/hooks';
-import {
-    TABLE_COMPACTION_OPERATION_PAGE_SIZE,
-    findRunningTableCompactionOperation,
-    isForcedCompactionEnabled,
-} from '../../../../../utils/tableCompaction';
-import {reachMetricaGoal} from '../../../../../utils/yaMetrica';
+import {isForcedCompactionEnabled} from '../../../../../utils/tableCompaction';
 import {EntityTitle} from '../../../EntityTitle/EntityTitle';
 import {isRowTableType} from '../../../utils/schema';
 
@@ -26,11 +17,12 @@ import {
     CompactTableAction,
     CompactTableStatusBanner,
 } from './CompactTableAction/CompactTableAction';
-import {openManagePartitioningDialog} from './ManagePartitioningDialog/ManagePartitioningDialog';
 import {PartitionsProgress} from './PartitionsProgress/PartitionsProgress';
+import {TableInfoSection} from './components/TableInfoSection';
+import {useTableCompaction} from './hooks/useTableCompaction';
+import {useTablePartitioning} from './hooks/useTablePartitioning';
 import i18n from './i18n';
 import {prepareTableInfo} from './prepareTableInfo';
-import {prepareUpdatePartitioningRequest} from './utils';
 
 import './TableInfo.scss';
 
@@ -56,9 +48,12 @@ const TableInfoHeader = ({data}: {data?: TEvDescribeSchemeResult}) => {
 };
 
 export const TableInfo = ({data, type, database, path}: TableInfoProps) => {
-    const [autoRefreshInterval] = useAutoRefreshInterval();
     const isRowTable = type === EPathType.EPathTypeTable;
     const featureFlagsAvailable = useFeatureFlagsAvailable();
+
+    // Prepare all table information
+    const tableInfo = React.useMemo(() => prepareTableInfo(data, type), [data, type]);
+
     const {
         generalInfoLeft = [],
         generalInfoRight = [],
@@ -69,72 +64,26 @@ export const TableInfo = ({data, type, database, path}: TableInfoProps) => {
         partitionConfigInfo = [],
         partitionProgressConfig,
         managePartitioningDialogConfig,
-    } = React.useMemo(() => prepareTableInfo(data, type), [data, type]);
+    } = tableInfo;
 
-    const [expanded, setExpanded] = React.useState(false);
-
-    const handleExpandedChange = React.useCallback((value: boolean) => setExpanded(value), []);
-
+    // Calculate if there's expandable content
     const hasMoreLeft = tableStatsInfo.some((items) => items.length > 0);
     const hasMoreRight = tabletMetricsInfo.length > 0 || partitionConfigInfo.length > 0;
-    const hasMore = hasMoreLeft || hasMoreRight;
 
-    const [updatePartitioning] = tablePartitioningApi.useUpdateTablePartitioningMutation();
+    // Compaction logic (only for row tables)
     const {currentData: featureFlags} = configsApi.useGetFeatureFlagsQuery(
         {database},
         {skip: !isRowTable || !featureFlagsAvailable},
     );
     const compactionEnabled = isRowTable && isForcedCompactionEnabled(featureFlags);
-    const {
-        currentData: compactionData,
-        isFetching: isCompactionFetching,
-        hasNextPage,
-        fetchNextPage,
-    } = operationsApi.useGetOperationListInfiniteQuery(
-        {database, kind: 'compaction', page_size: TABLE_COMPACTION_OPERATION_PAGE_SIZE},
-        {
-            pollingInterval: autoRefreshInterval,
-            skip: !compactionEnabled,
-        },
+    const {runningCompaction, isFetching: isCompactionFetching} = useTableCompaction(
+        database,
+        path,
+        compactionEnabled,
     );
 
-    // Automatically fetch all pages
-    React.useEffect(() => {
-        if (hasNextPage && !isCompactionFetching) {
-            fetchNextPage();
-        }
-    }, [hasNextPage, isCompactionFetching, fetchNextPage]);
-
-    const compactionOperations = React.useMemo(
-        () => compactionData?.pages?.flatMap((page) => page.operations ?? []),
-        [compactionData],
-    );
-    const runningCompaction = React.useMemo(
-        () => findRunningTableCompactionOperation(compactionOperations, path),
-        [compactionOperations, path],
-    );
     const [startTableCompaction, {isLoading: isCompactionStarting}] =
         operationsApi.useStartTableCompactionMutation();
-
-    const handleOpenManagePartitioning = React.useCallback(() => {
-        reachMetricaGoal('openManagePartitioning');
-        openManagePartitioningDialog({
-            initialValue: managePartitioningDialogConfig,
-            onApply: async (value) => {
-                reachMetricaGoal('applyManagePartitioning');
-                await updatePartitioning(
-                    prepareUpdatePartitioningRequest(value, database, path),
-                ).unwrap();
-
-                createToast({
-                    name: 'updateTablePartitioning',
-                    content: i18n('toast_partitioning-updated'),
-                    autoHiding: 3000,
-                    isClosable: true,
-                });
-            },
-        });
-    }, [managePartitioningDialogConfig, database, path, updatePartitioning]);
 
     const handleStartCompaction = React.useCallback(
         async ({cascade, parallel}: {cascade: boolean; parallel?: number}) => {
@@ -142,6 +91,18 @@ export const TableInfo = ({data, type, database, path}: TableInfoProps) => {
         },
         [database, path, startTableCompaction],
     );
+
+    // Partitioning logic
+    const {handleOpenManagePartitioning} = useTablePartitioning(
+        database,
+        path,
+        managePartitioningDialogConfig,
+    );
+
+    // UI state
+    const [expanded, setExpanded] = React.useState(false);
+    const handleExpandedChange = React.useCallback((value: boolean) => setExpanded(value), []);
+    const hasMore = React.useMemo(() => hasMoreLeft || hasMoreRight, [hasMoreLeft, hasMoreRight]);
 
     return (
         <div className={b()}>
@@ -191,50 +152,28 @@ export const TableInfo = ({data, type, database, path}: TableInfoProps) => {
             )}
             <div className={b('row', {'general-info': true})}>
                 <div className={b('col')}>
-                    {generalInfoLeft.length > 0 ? (
-                        <YDBDefinitionList
-                            nameMaxWidth="auto"
-                            responsive
-                            className={b('info-block')}
-                            items={generalInfoLeft}
-                        />
-                    ) : null}
+                    <TableInfoSection items={generalInfoLeft} className={b('info-block')} />
                 </div>
                 <div className={b('col')}>
-                    {generalInfoRight.length > 0 ? (
-                        <YDBDefinitionList
-                            nameMaxWidth="auto"
-                            responsive
-                            className={b('info-block')}
-                            items={generalInfoRight}
-                        />
-                    ) : null}
+                    <TableInfoSection items={generalInfoRight} className={b('info-block')} />
                 </div>
             </div>
 
             <div className={b('row')}>
                 <div className={b('col')}>
-                    {generalStats.length > 0 ? (
-                        <YDBDefinitionList
-                            items={generalStats}
-                            title={i18n('title_table-stats')}
-                            className={b('info-block')}
-                            nameMaxWidth="auto"
-                            responsive
-                        />
-                    ) : null}
+                    <TableInfoSection
+                        items={generalStats}
+                        title={i18n('title_table-stats')}
+                        className={b('info-block')}
+                    />
                 </div>
 
                 <div className={b('col')}>
-                    {generalMetrics.length > 0 ? (
-                        <YDBDefinitionList
-                            items={generalMetrics}
-                            title={i18n('title_tablet-metrics')}
-                            className={b('info-block')}
-                            nameMaxWidth="auto"
-                            responsive
-                        />
-                    ) : null}
+                    <TableInfoSection
+                        items={generalMetrics}
+                        title={i18n('title_tablet-metrics')}
+                        className={b('info-block')}
+                    />
                 </div>
             </div>
 
@@ -259,34 +198,24 @@ export const TableInfo = ({data, type, database, path}: TableInfoProps) => {
                                 {tableStatsInfo
                                     .filter((items) => items.length > 0)
                                     .map((items, index) => (
-                                        <YDBDefinitionList
+                                        <TableInfoSection
                                             key={index}
                                             items={items}
                                             className={b('info-block')}
-                                            nameMaxWidth="auto"
-                                            responsive
                                         />
                                     ))}
                             </div>
 
                             <div className={b('col')}>
-                                {tabletMetricsInfo.length > 0 ? (
-                                    <YDBDefinitionList
-                                        items={tabletMetricsInfo}
-                                        className={b('info-block')}
-                                        nameMaxWidth="auto"
-                                        responsive
-                                    />
-                                ) : null}
-                                {partitionConfigInfo.length > 0 ? (
-                                    <YDBDefinitionList
-                                        items={partitionConfigInfo}
-                                        title={i18n('title_partition-config')}
-                                        className={b('info-block')}
-                                        nameMaxWidth="auto"
-                                        responsive
-                                    />
-                                ) : null}
+                                <TableInfoSection
+                                    items={tabletMetricsInfo}
+                                    className={b('info-block')}
+                                />
+                                <TableInfoSection
+                                    items={partitionConfigInfo}
+                                    title={i18n('title_partition-config')}
+                                    className={b('info-block')}
+                                />
                             </div>
                         </div>
                     </Disclosure.Details>
